Move dsl under query package
