+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/metric/MetricJdbcReader.java
@@ -22,9 +22,9 @@ import org.bithon.server.commons.time.TimeSpan;
 import org.bithon.server.storage.datasource.DataSourceSchema;
 import org.bithon.server.storage.datasource.query.IQueryStageAggregator;
 import org.bithon.server.storage.datasource.query.QueryStageAggregators;
-import org.bithon.server.storage.jdbc.dsl.sql.NameExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.SelectExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.StringExpression;
+import org.bithon.server.storage.datasource.query.dsl.NameExpression;
+import org.bithon.server.storage.datasource.query.dsl.SelectExpression;
+import org.bithon.server.storage.datasource.query.dsl.StringExpression;
 import org.bithon.server.storage.jdbc.utils.SQLFilterBuilder;
 import org.bithon.server.storage.metrics.IFilter;
 import org.bithon.server.storage.metrics.IMetricReader;
