+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/metric/SelectExpressionBuilder.java
@@ -27,14 +27,14 @@ import org.bithon.server.storage.datasource.query.ast.FieldExpressionVisitorAdap
 import org.bithon.server.storage.datasource.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.spec.MetricSpecVisitorAdaptor;
 import org.bithon.server.storage.datasource.spec.PostAggregatorMetricSpec;
-import org.bithon.server.storage.jdbc.dsl.sql.GroupByExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.LimitExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.NameExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.OrderByExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.SelectExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.StringExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.TableExpression;
-import org.bithon.server.storage.jdbc.dsl.sql.WhereExpression;
+import org.bithon.server.storage.datasource.query.dsl.GroupByExpression;
+import org.bithon.server.storage.datasource.query.dsl.LimitExpression;
+import org.bithon.server.storage.datasource.query.dsl.NameExpression;
+import org.bithon.server.storage.datasource.query.dsl.OrderByExpression;
+import org.bithon.server.storage.datasource.query.dsl.SelectExpression;
+import org.bithon.server.storage.datasource.query.dsl.StringExpression;
+import org.bithon.server.storage.datasource.query.dsl.TableExpression;
+import org.bithon.server.storage.datasource.query.dsl.WhereExpression;
 import org.bithon.server.storage.jdbc.utils.SQLFilterBuilder;
 import org.bithon.server.storage.metrics.IFilter;
 import org.bithon.server.storage.metrics.Interval;
