+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/dsl/FieldsExpression.java
@@ -14,21 +14,39 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.dsl.sql;
+package org.bithon.server.storage.datasource.query.dsl;
 
-import lombok.Data;
+import lombok.Getter;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2022/9/4 15:43
+ * @date 2022/9/4 14:56
  */
-@Data
-public class FromExpression implements IExpression {
-    private IExpression expression;
+public class FieldsExpression implements IExpression {
+    @Getter
+    private final List<IExpression> fields = new ArrayList<>(8);
+
+    public FieldsExpression insert(IExpression field) {
+        fields.add(0, field);
+        return this;
+    }
+
+    public FieldsExpression addField(IExpression field) {
+        fields.add(field);
+        return this;
+    }
+
+    public void addFields(List<String> fields) {
+        for (String field : fields) {
+            this.fields.add(new NameExpression(field));
+        }
+    }
 
     @Override
     public void accept(IExpressionVisitor visitor) {
         visitor.visit(this);
-        expression.accept(visitor);
     }
 }
