+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/dsl/FromExpression.java
@@ -14,32 +14,21 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.dsl.sql;
+package org.bithon.server.storage.datasource.query.dsl;
 
-import lombok.Getter;
-
-import java.util.ArrayList;
-import java.util.List;
+import lombok.Data;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2022/9/4 16:40
+ * @date 2022/9/4 15:43
  */
-@Getter
-public class FunctionExpression implements IExpression {
-    private final String fnName;
-    private final List<IExpression> arguments = new ArrayList<>();
-
-    public FunctionExpression(String fnName) {
-        this.fnName = fnName;
-    }
+@Data
+public class FromExpression implements IExpression {
+    private IExpression expression;
 
     @Override
     public void accept(IExpressionVisitor visitor) {
-        visitor.before(this);
-        for (IExpression arg : arguments) {
-            arg.accept(visitor);
-        }
-        visitor.after(this);
+        visitor.visit(this);
+        expression.accept(visitor);
     }
 }
