+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/dsl/IExpressionVisitor.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.dsl.sql;
+package org.bithon.server.storage.datasource.query.dsl;
 
 /**
  * @author frank.chen021@outlook.com
