+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/dsl/NameExpression.java
@@ -14,15 +14,21 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.dsl.sql;
+package org.bithon.server.storage.datasource.query.dsl;
+
+import lombok.Getter;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2022/9/4 16:34
+ * @date 2022/9/4 16:11
  */
-public class AliasExpression extends NameExpression {
-    public AliasExpression(String name) {
-        super(name);
+@Getter
+public class NameExpression implements IExpression {
+
+    private final String name;
+
+    public NameExpression(String name) {
+        this.name = name;
     }
 
     @Override
