+++ server/storage/src/main/java/org/bithon/server/storage/datasource/query/dsl/SelectExpression.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.dsl.sql;
+package org.bithon.server.storage.datasource.query.dsl;
 
 import lombok.Data;
 import lombok.Getter;
